fix: rename razorpay config file
